修复登录问题，简化模型
