+++ src/Pixeval/Controls/IllustratorItem/IllustratorItemViewModel.Commands.cs
@@ -1,14 +1,8 @@
 using System;
-using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Input;
 using Pixeval.Controls.MarkupExtensions;
 using Pixeval.Util;
-using Pixeval.Util.IO;
 using Pixeval.Util.UI;
-using WinUI3Utilities;
-using Microsoft.UI.Xaml.Media.Imaging;
-using Windows.System;
 
 namespace Pixeval.Controls;
 
@@ -16,30 +10,12 @@ public partial class IllustratorItemViewModel
 {
     public XamlUICommand FollowCommand { get; } = "".GetCommand(FontIconSymbol.ContactE77B);
 
-    public XamlUICommand GenerateLinkCommand { get; } = EntryItemResources.GenerateLink.GetCommand(FontIconSymbol.LinkE71B);
-
-    public XamlUICommand GenerateWebLinkCommand { get; } = EntryItemResources.GenerateWebLink.GetCommand(FontIconSymbol.PreviewLinkE8A1);
-
-    public XamlUICommand OpenInWebBrowserCommand { get; } = EntryItemResources.OpenInWebBrowser.GetCommand(FontIconSymbol.WebSearchF6FA);
-
-    public XamlUICommand ShowQrCodeCommand { get; } = EntryItemResources.ShowQRCode.GetCommand(FontIconSymbol.QRCodeED14);
-
-    public XamlUICommand ShowPixEzQrCodeCommand { get; } = EntryItemResources.ShowPixEzQrCode.GetCommand(FontIconSymbol.Photo2EB9F);
-
     private void InitializeCommands()
     {
+        InitializeCommandsBase();
+
         FollowCommand.GetFollowCommand(IsFollowed);
         FollowCommand.ExecuteRequested += FollowCommandExecuteRequested;
-
-        GenerateLinkCommand.ExecuteRequested += GenerateLinkCommandOnExecuteRequested;
-
-        GenerateWebLinkCommand.ExecuteRequested += GenerateWebLinkCommandOnExecuteRequested;
-
-        OpenInWebBrowserCommand.ExecuteRequested += OpenInWebBrowserCommandOnExecuteRequested;
-
-        ShowQrCodeCommand.ExecuteRequested += ShowQrCodeCommandExecuteRequested;
-
-        ShowPixEzQrCodeCommand.ExecuteRequested += ShowPixEzQrCodeCommandExecuteRequested;
     }
 
     private async void FollowCommandExecuteRequested(XamlUICommand sender, ExecuteRequestedEventArgs args)
@@ -48,58 +24,9 @@ public partial class IllustratorItemViewModel
         FollowCommand.GetFollowCommand(IsFollowed);
     }
 
-    private void GenerateLinkCommandOnExecuteRequested(XamlUICommand sender, ExecuteRequestedEventArgs args)
-    {
-        UiHelper.ClipboardSetText(MakoHelper.GenerateIllustrationAppUri(UserId).OriginalString);
+    protected override Uri AppUri => MakoHelper.GenerateUserAppUri(UserId);
 
-        if (args.Parameter is not TeachingTip teachingTip)
-            return;
-        if (App.AppViewModel.AppSettings.DisplayTeachingTipWhenGeneratingAppLink)
-            teachingTip.IsOpen = true;
-        else
-            teachingTip?.ShowTeachingTipAndHide(EntryItemResources.LinkCopiedToClipboard);
-    }
+    protected override Uri WebUri => MakoHelper.GenerateUserWebUri(UserId);
 
-    private void GenerateWebLinkCommandOnExecuteRequested(XamlUICommand sender, ExecuteRequestedEventArgs args)
-    {
-        UiHelper.ClipboardSetText(MakoHelper.GenerateUserWebUri(UserId).OriginalString);
-        (args.Parameter as FrameworkElement)?.ShowTeachingTipAndHide(EntryItemResources.LinkCopiedToClipboard);
-    }
-
-    private async void OpenInWebBrowserCommandOnExecuteRequested(XamlUICommand sender, ExecuteRequestedEventArgs args)
-    {
-        _ = await Launcher.LaunchUriAsync(MakoHelper.GenerateUserWebUri(UserId));
-    }
-
-    private async void ShowQrCodeCommandExecuteRequested(XamlUICommand sender, ExecuteRequestedEventArgs args)
-    {
-        if (args.Parameter is not TeachingTip showQrCodeTeachingTip)
-            return;
-
-        var qrCodeSource = await IoHelper.GenerateQrCodeForUrlAsync(MakoHelper.GenerateUserWebUri(UserId).OriginalString);
-        ShowQrCodeCommandExecuteRequested(showQrCodeTeachingTip, qrCodeSource);
-    }
-
-    private async void ShowPixEzQrCodeCommandExecuteRequested(XamlUICommand sender, ExecuteRequestedEventArgs args)
-    {
-        if (args.Parameter is not TeachingTip showQrCodeTeachingTip)
-            return;
-
-        var qrCodeSource = await IoHelper.GenerateQrCodeAsync(MakoHelper.GenerateUserPixEzUri(UserId).OriginalString);
-        ShowQrCodeCommandExecuteRequested(showQrCodeTeachingTip, qrCodeSource);
-    }
-
-    private static void ShowQrCodeCommandExecuteRequested(TeachingTip teachingTip, SoftwareBitmapSource source)
-    {
-        teachingTip.HeroContent.To<Image>().Source = source;
-        teachingTip.IsOpen = true;
-        teachingTip.Closed += Closed;
-        return;
-
-        void Closed(TeachingTip s, TeachingTipClosedEventArgs ea)
-        {
-            source.Dispose();
-            s.Closed -= Closed;
-        }
-    }
+    protected override Uri PixEzUri => MakoHelper.GenerateUserPixEzUri(UserId);
 }
